requirements.txt
fastapi==0.104.1
httpx==0.25.1
langchain==0.0.316
openai==0.28.1
python-dotenv==1.0.0
